let users define names of options files
This is a hack to select build variants, e.g. debug vs. releasem
by using different options files named via environment variables.

The current approach is fine grained, one variable per file name.
This could perhaps be simplified to a single variable to define
a "build variant" that gets encoded into the name somehow. The
current approach is intended to be used by programs (dmake) so
its not too bad. The flexibility is nice and perhaps should
be retained and a simpler method added.